fix: method on request
