+++ scripts/router.ts
@@ -1,6 +1,6 @@
 import { advancedEval } from "@utils/advancedEval";
-import type { Route, RouteBuildedFunction } from "./duplose/route";
-import type { HttpMethod, RequestInitializationData } from "./request";
+import type { HttpMethod, Route, RouteBuildedFunction } from "./duplose/route";
+import type { RequestInitializationData } from "./request";
 import { getTypedEntries } from "@utils/getTypedEntries";
 import { StringBuilder } from "@utils/stringBuilder";
 import { hasKey } from "@utils/hasKey";
