+++ test/utils/duploTest.ts
@@ -1,12 +1,16 @@
-import { Duplo, Route, Router, type PromiseOrNot } from "@scripts/index";
+import { Duplo, Route, Router, useRouteBuilder, type PromiseOrNot } from "@scripts/index";
 
 export class DuploTest extends Duplo {
 	public async start(onStart?: (duplo: Duplo) => PromiseOrNot<void>) {
+		const notfoundHandler = this.notfoundHandler;
+		const notfoundRoute = useRouteBuilder(new Route("GET", ["/*"])).handler((pickup, request) => notfoundHandler(request));
+		this.register(notfoundRoute);
+
 		await this.hooksInstanceLifeCycle.beforeBuildRouter.launchSubscriber(this);
 
 		const router = new Router(
 			this.duploses.filter((duplose) => duplose instanceof Route),
-			this.createNotfoundRoute(),
+			notfoundRoute,
 		);
 
 		if (onStart) {
